requirements.txt
Django==1.8.9
django-bootstrap3==6.2.2
django-localflavor==1.2
django-simple-menu==1.2.0
